refactor: add ChatMessageContent
