+++ src/pages/inbox/report/ReportActionItem.tsx
@@ -1,7 +1,5 @@
-import {validTransactionDraftIDsSelector} from '@selectors/TransactionDraft';
 import React, {useCallback} from 'react';
 import type {OnyxEntry} from 'react-native-onyx';
-import {useBlockedFromConcierge} from '@components/OnyxListItemProvider';
 import useCurrentUserPersonalDetails from '@hooks/useCurrentUserPersonalDetails';
 import useLocalize from '@hooks/useLocalize';
 import useOnyx from '@hooks/useOnyx';
@@ -14,7 +12,6 @@ import {getForReportAction, getMovedReportID} from '@libs/ModifiedExpenseMessage
 import {getIOUReportIDFromReportActionPreview, getOriginalMessage, isMoneyRequestAction} from '@libs/ReportActionsUtils';
 import {
     chatIncludesChronosWithID,
-    createDraftTransactionAndNavigateToParticipantSelector,
     getIndicatedMissingPaymentMethod,
     getReimbursementDeQueuedOrCanceledActionMessage,
     getTransactionsWithReceipts,
@@ -24,13 +21,7 @@ import {
     isCurrentUserTheOnlyParticipant,
 } from '@libs/ReportUtils';
 import {clearAllRelatedReportActionErrors} from '@userActions/ClearReportActionErrors';
-import {
-    deleteReportActionDraft,
-    dismissTrackExpenseActionableWhisper,
-    resolveActionableMentionWhisper,
-    resolveActionableReportMentionWhisper,
-    toggleEmojiReaction,
-} from '@userActions/Report';
+import {deleteReportActionDraft, resolveActionableMentionWhisper, resolveActionableReportMentionWhisper, toggleEmojiReaction} from '@userActions/Report';
 import {clearError} from '@userActions/Transaction';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
@@ -38,18 +29,7 @@ import type {PersonalDetailsList, ReportAction, ReportActionReactions, Transacti
 import type {PureReportActionItemProps} from './PureReportActionItem';
 import PureReportActionItem from './PureReportActionItem';
 
-type ReportActionItemProps = Omit<
-    PureReportActionItemProps,
-    | 'taskReport'
-    | 'linkedReport'
-    | 'iouReportOfLinkedReport'
-    | 'currentUserAccountID'
-    | 'currentUserEmail'
-    | 'personalPolicyID'
-    | 'draftTransactionIDs'
-    | 'userBillingGracePeriodEnds'
-    | 'betas'
-> & {
+type ReportActionItemProps = Omit<PureReportActionItemProps, 'taskReport' | 'linkedReport' | 'iouReportOfLinkedReport' | 'currentUserAccountID' | 'personalPolicyID' | 'betas'> & {
     /** Whether to show the draft message or not */
     shouldShowDraftMessage?: boolean;
 
@@ -103,7 +83,6 @@ function ReportActionItem({
     const [policyTags] = useOnyx(`${ONYXKEYS.COLLECTION.POLICY_TAGS}${policyIDForTags}`);
     const [introSelected] = useOnyx(ONYXKEYS.NVP_INTRO_SELECTED);
     const [betas] = useOnyx(ONYXKEYS.BETAS);
-    const [draftTransactionIDs] = useOnyx(ONYXKEYS.COLLECTION.TRANSACTION_DRAFT, {selector: validTransactionDraftIDsSelector});
     const [reportMetadata] = useOnyx(`${ONYXKEYS.COLLECTION.REPORT_METADATA}${reportID}`);
     const [originalReport] = useOnyx(`${ONYXKEYS.COLLECTION.REPORT}${originalReportID}`);
     const [iouReport] = useOnyx(`${ONYXKEYS.COLLECTION.REPORT}${getIOUReportIDFromReportActionPreview(action)}`);
@@ -122,7 +101,6 @@ function ReportActionItem({
     const [cardList] = useOnyx(ONYXKEYS.CARD_LIST);
     const [bankAccountList] = useOnyx(ONYXKEYS.BANK_ACCOUNT_LIST);
     const [personalPolicyID] = useOnyx(ONYXKEYS.PERSONAL_POLICY_ID);
-    const [userBillingGracePeriodEnds] = useOnyx(ONYXKEYS.COLLECTION.SHARED_NVP_PRIVATE_USER_BILLING_GRACE_PERIOD_END);
     const transactionsOnIOUReport = useReportTransactions(iouReport?.reportID);
     const transactionID = isMoneyRequestAction(action) && getOriginalMessage(action)?.IOUTransactionID;
 
@@ -135,7 +113,6 @@ function ReportActionItem({
 
     const [linkedTransactionRouteError] = useOnyx(`${ONYXKEYS.COLLECTION.TRANSACTION}${transactionID}`, {selector: getLinkedTransactionRouteError});
 
-    const blockedFromConcierge = useBlockedFromConcierge();
     const targetReport = isChatThread(report) ? parentReport : report;
     const missingPaymentMethod = getIndicatedMissingPaymentMethod(userWalletTierName, targetReport?.reportID, action, bankAccountList);
 
@@ -145,13 +122,11 @@ function ReportActionItem({
             {...props}
             introSelected={introSelected}
             betas={betas}
-            draftTransactionIDs={draftTransactionIDs}
             personalPolicyID={personalPolicyID}
             action={action}
             report={report}
             policy={policy}
             currentUserAccountID={currentUserAccountID}
-            currentUserEmail={currentUserEmail}
             draftMessage={draftMessage}
             iouReport={iouReport}
             taskReport={taskReport}
@@ -163,14 +138,12 @@ function ReportActionItem({
             isUserValidated={isUserValidated}
             parentReport={parentReport}
             personalDetails={personalDetails}
-            blockedFromConcierge={blockedFromConcierge}
             originalReportID={originalReportID}
             originalReport={originalReport}
             deleteReportActionDraft={deleteReportActionDraft}
             isArchivedRoom={isArchivedNonExpenseReport(originalReport, isOriginalReportArchived)}
             isChronosReport={chatIncludesChronosWithID(originalReportID)}
             toggleEmojiReaction={toggleEmojiReaction}
-            createDraftTransactionAndNavigateToParticipantSelector={createDraftTransactionAndNavigateToParticipantSelector}
             resolveActionableReportMentionWhisper={resolveActionableReportMentionWhisper}
             resolveActionableMentionWhisper={resolveActionableMentionWhisper}
             isClosedExpenseReportWithNoExpenses={isClosedExpenseReportWithNoExpenses(iouReport, transactionsOnIOUReport)}
@@ -193,12 +166,10 @@ function ReportActionItem({
             getTransactionsWithReceipts={getTransactionsWithReceipts}
             clearError={clearError}
             clearAllRelatedReportActionErrors={clearAllRelatedReportActionErrors}
-            dismissTrackExpenseActionableWhisper={dismissTrackExpenseActionableWhisper}
             userBillingFundID={userBillingFundID}
             isTryNewDotNVPDismissed={isTryNewDotNVPDismissed}
             bankAccountList={bankAccountList}
             reportMetadata={reportMetadata}
-            userBillingGracePeriodEnds={userBillingGracePeriodEnds}
         />
     );
 }
